refactor : enum을 constants 패키지로 이동시킨다.
