--- src/main/java/subway/view/Sentence.java
@@ -1,13 +1,0 @@
-package subway.view;
-
-public enum Sentence {
-
-    PREFIX("## "),
-    INFO("[INFO] ");
-
-    final String message;
-
-    Sentence(String message) {
-        this.message = message;
-    }
-}
